main/backend/requirements.txt
Flask==2.3.2
Flask-Cors==4.0.0
pyjwt==2.8.0
python-dotenv==1.0.1
openai==1.35.14
Authlib==1.2.0
requests==2.28.0
